fix(mockData): 임시 카드 데이터 경로 오류 수정
다른 위치로 이동해서 문제 해결
